backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
redis
python-jose[cryptography]
passlib[bcrypt]
pydantic
alembic
python-multipart
Jinja2
requests
